Added python files
Akhinator is a gaming application designed using python. This gaming application will provide the user a list of fictional or real-life personalities where,the user can challenge the application to guess the personality that they had selected by answering questions posed by the application. (The actual version of akhinator does not confine the users' list of choices). 

HAVI - This is a very basic  Ecommerce project implemented using python, which facilitates shopping of watches online. This application provides a wide range of watches under different categories according to the needs of the customers. It facilitates immediate purchases and also provides cart functionalities. (It does not connect to a payment gateway).